+++ Chromagnon/Priithon/viewerCommon.py
@@ -49,7 +49,7 @@ Menu_noGfx = wx.NewId()
 Menu_aspectRatio = wx.NewId()
 Menu_rotate = wx.NewId()
 Menu_grid        = wx.NewId()
-Menu_ColMap = [wx.NewId() for i in range(8)]
+Menu_ColMap = [wx.NewId() for i in range(9)]
 
 
 class GLViewerCommon(wxgl.GLCanvas):