+++ Chromagnon/ndviewer/viewerCommon.py
@@ -583,7 +583,10 @@ class GLViewerCommon(wxgl.GLCanvas):
     ## don't need
     def OnSaveScreenShort(self, event=None):
         '''always flipY'''
-        from Priithon.all import U, FN
+        try:
+            from ..Priithon.all import U, FN
+        except (ValueError, ImportError):
+            from Priithon.all import U, FN
         fn = FN(1)#, verbose=0)
         if not fn:
             return